Rename inject-encoding-test to custom-encoding-test
