+++ test/custom-encoding-test.js
