+++ test/index.js
@@ -31,7 +31,7 @@ require('./deferred-open-test')
 require('./get-put-del-test')
 require('./idempotent-test')(test, testCommon)
 require('./init-test')
-require('./inject-encoding-test')(test, testCommon)
+require('./custom-encoding-test')(test, testCommon)
 require('./json-test')
 require('./key-value-streams-test')
 require('./maybe-error-test')(test, testCommon)
